Moved the stub coverage tests into a new file
